leapfrog + slightplotting changes Leapfrog neccecitates an architecture change.
Right now, what we do is
move particle A, keep it's old position
change particle A's velocity,
move particle B.
etc.

For leapfrog, what we need to do is
Move A,
Move B,
...
Change velocity of A
change vel. of B

This can **not** be done in the Atom class. It needs to be done in Particles.